feat: publish package, create wrapper and create example
