+++ sdk/README.md
@@ -1,0 +1,10 @@
+# Testkube Typescript SDK 
+
+Typescript SDK for running Testkube tests and creating pipelines.
+
+The Types and underlying client are generated from Testkube's OpenAPI definition: 
+
+```sh 
+npx oazapfts ../api/v1/testkube.yaml src/generated-client.ts
+```
+
